Prevent date picker scrolling the whole page when focused
